+++ src/bouncer/core.clj
@@ -1,11 +1,11 @@
 (ns bouncer.core
-  "The `core` namespace provides the two main validation macros in bouncer:
+  "The `core` namespace provides the two main entry point functions in bouncer:
 
   - `validate`
-  - `valid`
+  - `valid?`
 
 
-All other functions are meant for internal use only and shouldn't be relied on.
+All other functions are meant for internal use only and shouldn't be relied upon.
 
 The project [README](https://github.com/leonardoborges/bouncer/blob/master/README.md) should get you started,
 it's pretty comprehensive.
@@ -14,86 +14,81 @@ it's pretty comprehensive.
 If you'd like to know more about the motivation behind `bouncer`, check the
 [announcement post](http://www.leonardoborges.com/writings/2013/01/04/bouncer-validation-lib-for-clojure/)."
   {:author "Leonardo Borges"}
-  (:require [clojure.algo.monads :as m]
-            [bouncer.helpers :as h]))
+  (:require [clojure.algo.monads :as m]))
 
 
 
 ;; ## Internal utility functions
 
-(defn build-multi-step
+(defn- build-multi-step
   ([key-or-vec fn-vec] (build-multi-step key-or-vec fn-vec []))
   ([key-or-vec [f-or-list & rest] acc]
      (if-not f-or-list
        acc
        (cond
-        (list? f-or-list)
+        (sequential? f-or-list)
         (let [[f & opts] f-or-list]
           (recur key-or-vec
                  rest
-                 (conj acc `[~(h/resolve-or-same f) ~key-or-vec ~@opts])))
+                 (conj acc (concat [f key-or-vec ] opts))))
 
         :else (recur key-or-vec
                      rest
-                     (conj acc `[~(h/resolve-or-same f-or-list) ~key-or-vec]))))))
+                     (conj acc [f-or-list key-or-vec]))))))
 
-(def is-validator-set? (comp true?
-                             :bouncer-validator-set
-                             meta
-                             h/resolve-or-same))
-
-(defn merge-path
+(defn- merge-path
   "Takes two arguments:
 
-  `parent-keyword` is a a keyword - or a vector of :keywords denoting a path in a associative structure
+  `parent-keyword` is a :keyword - or a vector of :keywords denoting a path in a associative structure
 
-  `coll` is a seq of forms following this spec:
-
-
-      (:keyword [f g] :another-keyword h)
+  `validations-map` is a map of forms following this spec:
 
 
-  Merges `:parent-keyword` with every first element of coll, transforming coll into:
+      {:keyword [f g] :another-keyword h}
+
+
+  Merges `:parent-keyword` with every first element of validations-map, transforming it into:
 
 
       ([:parent-keyword :keyword] [f g] [:parent-keyword :another-keyword] h)
 "
-  [parent-key coll]
-  (let [parent-key (if (keyword? parent-key) [parent-key] parent-key)
-        pairs (partition 2 coll)]
-    (mapcat #(if (vector? (first %))
-            [(apply vector (concat parent-key (first %))) (second %)]
-            [(apply vector (concat parent-key [(first %)])) (second %)])
-    pairs)))
+  [parent-key validations-map]
+  (let [parent-key (if (keyword? parent-key) [parent-key] parent-key)]
+    (mapcat (fn [[key validations]]
+              (if (vector? key)
+                [(apply vector (concat parent-key key)) validations]
+                [(apply vector (concat parent-key [key])) validations]))
+            validations-map)))
 
-(defn build-steps [forms]
-  (reduce (fn [acc [key-or-vec sym-or-coll :as rule]]
-            (cond
-             (vector? sym-or-coll)
-             (concat acc (build-multi-step key-or-vec sym-or-coll))
-             
-             (list? sym-or-coll)
-             (concat acc (build-multi-step key-or-vec [sym-or-coll]))
-             
-             (is-validator-set? sym-or-coll)
-             (concat acc (build-steps (merge-path key-or-vec
-                                                  (var-get (h/resolve-or-same sym-or-coll)))))
+(defn- build-steps [[head & tail :as forms]]  
+  (let [forms (if (map? head)
+                (vec (mapcat identity head))
+                forms)]
+    (reduce (fn [acc [key-or-vec sym-or-coll :as rule]]
+              (cond
+               (vector? sym-or-coll)
+               (concat acc (build-multi-step key-or-vec sym-or-coll))
+               
+               
+               (map? sym-or-coll)
+               (concat acc (build-steps (merge-path key-or-vec
+                                                    sym-or-coll)))
 
-             :else (conj acc `[~(h/resolve-or-same sym-or-coll) ~key-or-vec])))
-          []
-          (partition 2 forms)))
+               :else (conj acc [sym-or-coll key-or-vec])))
+            []
+            (partition 2 forms))))
 
-(defn pre-condition-met? [pre-fn map]
+(defn- pre-condition-met? [pre-fn map]
   (or (nil? pre-fn) (pre-fn map)))
 
-(defn wrap
+(defn- wrap
   "Wraps pred in the context of validating a single value
 
   - `acc`  is the map being validated
 
   - `pred` is a validator
 
-  - `k`    the path to the value to be validated in the associative structure acc
+  - `k`    the path to the value to be validated in the associative structure `acc`
 
   - `args` any extra args to pred
 
@@ -106,12 +101,12 @@ If you'd like to know more about the motivation behind `bouncer`, check the
   Returns `acc` augmented with a namespace qualified ::errors keyword
 "
   [acc [pred k & args]]
-  (let [pred (h/resolve-or-same pred)
-        k (if (vector? k) k [k])
+  (let [k (if (vector? k) k [k])
         error-path (cons ::errors k)
-        {:keys [default-message-format optional]} (meta pred)
+        {:keys [default-message-format optional]
+         :or {default-message-format "Custom validation failed for %s"
+              optional false}} (meta pred)
         [args opts] (split-with (complement keyword?) args)
-        args (map h/get-var-or-same args)
         {:keys [message pre] :or {message default-message-format}} (apply hash-map opts)
         pred-subject (get-in acc k)]
     (if (pre-condition-met? pre acc)
@@ -123,7 +118,8 @@ If you'd like to know more about the motivation behind `bouncer`, check the
                    #(conj % (format message (name (peek k))))))
       acc)))
 
-(defn wrap-chain
+
+(defn- wrap-chain
   "Internal Use.
 
   Chain of responsibility.
@@ -134,34 +130,35 @@ If you'd like to know more about the motivation behind `bouncer`, check the
 
   See also `wrap`
 "
-  [fs]
+  [& fs]
   (fn [old-state]
     (let [new-state (reduce wrap
                             old-state
                             fs)]
       [(::errors new-state) new-state])))
 
-(defn emit-wrap [entry]
-  `(wrap-chain ~(second entry)))
-
-(defmacro validate*
+(defn- validate*
   "Internal use.
 
   Validates the map m using the validation functions fs.
 
   Returns a vector where the first element is the map of validation errors if any and the second is the original map (possibly)augmented with the errors map."
-  [m & fs]
-  (let [ignore (gensym "ignore__")
-        result (gensym "result__")
-        wrap-calls (map emit-wrap fs)
-        step-pairs (vec (interleave (repeat ignore) wrap-calls))]
-    `((m/domonad m/state-m
-                 ~(assoc step-pairs (- (count step-pairs) 2) result)
-                 ~result) ~m)))
+  [m fs]
+  (letfn [(m-fn [fs]
+            (let [m-bind (:m-bind m/state-m)
+                  m-result (:m-result m/state-m)]
+              (cond
+               (> (count fs) 1) (m-bind (bouncer.core/wrap-chain (first fs))
+                                        (fn [_]
+                                          (m-fn (rest fs))))
+               :else (m-bind (bouncer.core/wrap-chain (first fs))
+                             (fn [result]
+                               (m-result result))))))]
+    ((m-fn fs) m)))
 
 ;; ## Public API
 
-(defmacro validate
+(defn validate
   "Validates the map m using the validations specified by forms.
 
   forms can be a single validator set or a sequence of key/value pairs where:
@@ -170,31 +167,27 @@ If you'd like to know more about the motivation behind `bouncer`, check the
 
   value ==> validation-function or
             validator-set or
-           (validation-function args+opts) or
+           [[validation-function args+opts]] or
            [validation-function another-validation-function] or
-           [validation-function (another-validation-function args+opts)]
+           [validation-function [another-validation-function args+opts]]
 
   e.g.:
 
       (core/validate a-map
                :name core/required
-               :age [core/required
-                     (core/number :message \"age must be a number\")]
+               :age  [core/required
+                     [core/number :message \"age must be a number\"]]
                [:passport :number] core/positive)
 
 
   Returns a vector where the first element is the map of validation errors if any and the second is the original map (possibly)augmented with the errors map.
 
-  See also `defvalidatorset`
+  See also `defvalidator`
 "
   [m & forms]
-  (if (= (count forms) 1)
-    `(validate* ~m
-                ~@(group-by second (build-steps (var-get (h/resolve-or-same (first forms))))))
-    `(validate* ~m
-                ~@(group-by second (build-steps forms)))))
+  (validate* m (build-steps forms)))
 
-(defmacro valid?
+(defn valid?
   "Takes a map and one or more validation functions with semantics provided by \"validate\". Returns true if the map passes all validations. False otherwise."
   [& args]
-  `(empty? (first (validate ~@args))))
+  (empty? (first (apply validate args))))