+++ src/bouncer/validators.clj
@@ -2,8 +2,7 @@
   "This namespace contains all built-in validators as well as
           macros for defining new validators and validator sets"
   {:author "Leonardo Borges"}
-  (:require [clojure.walk :as w]
-            [bouncer.helpers :as h]))
+  (:require [clojure.walk :as w]))
 
 ;; ## Customization support
 ;;
@@ -32,7 +31,7 @@
       (some #{value} coll))
 
     (validate {:age 10}
-      :age (in (range 5)))
+      :age [[member (range 5)]])
 
 
   This means the validator `member` will be called with the arguments `10` and `(0 1 2 3 4)`, 
@@ -58,37 +57,17 @@
                   (next options)
                   options)
         [args & body] options
-        fn-meta {:doc docstring
-                 :default-message-format default-message-format
+        fn-meta {:default-message-format default-message-format
                  :optional optional}]
-    `(defn ~(with-meta name fn-meta)
-       {:arglists '([~@args])}
-       ([~@args]
-          ~@body))))
-
-;; ### Composability
-
-
-(defmacro defvalidatorset
-  "Defines a set of validators encapsulating a reusable validation unit.
-
-  forms should follow the semantics of \"bouncer.core/validate\"
-
-  e.g.:
-
-    (defvalidatorset addr-validator-set
-      :postcode  [v/required v/number]
-      :street    v/required
-      :country   v/required)
-
-    (defvalidatorset person-validator-set
-      :name  [v/required v/number]
-      :address addr-validator-set)
-"
-  [name & forms]
-  `(def ~(with-meta name {:bouncer-validator-set true})
-     '(~@(w/postwalk h/resolve-or-same forms))))
-
+    (let [arglists ''([name])]
+      `(do (def ~name
+             (with-meta (fn ~name 
+                          ([~@args]
+                             ~@body))
+               (merge ~fn-meta)))
+           (alter-meta! (var ~name) assoc
+                        :doc ~docstring
+                        :arglists '([~@args]))))))
 
 ;; ## Built-in validators
 
@@ -97,7 +76,7 @@
 
   If the value is a string, it makes sure it's not empty, otherwise it checks for nils.
 
-  For use with validation macros such as `validate` or `valid?`
+  For use with validation functions such as `validate` or `valid?`
 "
   {:default-message-format "%s must be present"}
   [value]
@@ -108,7 +87,7 @@
 (defvalidator number
   "Validates maybe-a-number is a valid number.
 
-  For use with validation macros such as `validate` or `valid?`"
+  For use with validation functions such as `validate` or `valid?`"
   {:default-message-format "%s must be a number" :optional true}
   [maybe-a-number]
   (number? maybe-a-number))
@@ -117,7 +96,7 @@
 (defvalidator positive
   "Validates number is a number and is greater than zero.
 
-  For use with validation macros such as `validate` or `valid?`"
+  For use with validation functions such as `validate` or `valid?`"
   {:default-message-format "%s must be a positive number" :optional true}
   [number]
   (> number 0))
@@ -126,7 +105,7 @@
 (defvalidator member
   "Validates value is a member of coll.
 
-  For use with validation macros such as `validate` or `valid?`"
+  For use with validation functions such as `validate` or `valid?`"
   {:default-message-format "%s must be one of the values in the list"}
   [value coll]
   (some #{value} coll))
@@ -134,14 +113,15 @@
 (defvalidator custom
   "Validates pred is true for the given value.
 
-  For use with validation macros such as `validate` or `valid?`"
+  For use with validation functions such as `validate` or `valid?`"
   [value pred]
+  (println "Warning: bouncer.validators/custom is deprecated and will be removed. Use plain functions instead.")
   (pred value))
 
 (defvalidator every
   "Validates pred is true for every item in coll.
 
-  For use with validation macros such as `validate` or `valid?`"
+  For use with validation functions such as `validate` or `valid?`"
   {:default-message-format "All items in %s must satisfy the predicate"}
   [coll pred]
   (every? pred coll))
@@ -149,7 +129,7 @@
 (defvalidator matches
   "Validates value satisfies the given regex pattern.
 
-   For use with validation macros such as `validate` or `valid?`"
+   For use with validation functions such as `validate` or `valid?`"
   {:default-message-format "%s must satisfy the given pattern" :optional true}
   [value re]
   ((complement empty?) (re-seq re value)))